split feedback modal into seperate components
